backendraif2/requirements.txt
# Dr. Raif 2 - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
openai==1.10.0

# RAG & Embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.26.3

# Utils
python-multipart==0.0.6
